Refactored ThreadPool into a separate module
