+++ rust-web-server/src/lib.rs
@@ -1,9 +1,10 @@
+pub mod multithreading;
+
 use std::{
     fs,
     io::{prelude::*, BufReader},
     net::TcpStream,
-    sync::{mpsc, Arc, Mutex},
-    thread,
+
 };
 
 pub fn handle_connection(mut stream: TcpStream) {
@@ -31,82 +32,6 @@ fn parse_http_request(request: &String) -> (&str, &str) {
     match request.split_whitespace().collect::<Vec<&str>>().as_slice() {
         ["GET", "/", _] => ("HTTP/1.1 200 OK", "hello.html"),
         _ => ("HTTP/1.1 404 NOT FOUND", "404.html"),
-    }
-}
-
-type Job = Box<dyn FnOnce() + Send + 'static>;
-
-pub struct ThreadPool{
-    workers: Vec<Worker>,
-    dispatcher: Option<mpsc::Sender<Job>>,
-}
-
-impl ThreadPool{
-    /// Create a new ThreadPool.
-    /// 
-    /// The size is the number of threads in the pool.
-    /// 
-    /// # Panics
-    /// 
-    /// The `new` function will panic if the size is zero.
-    pub fn new(size: usize) -> ThreadPool {
-        assert!(size > 0);
-        let mut workers = Vec::with_capacity(size as usize);
-
-        let (dispatcher, receiver) = mpsc::channel();
-
-        let receiver = Arc::new(Mutex::new(receiver));
-
-        for id in 0..size {
-            // create some threads and store them in the vector
-            workers.push(Worker::new(id, Arc::clone(&receiver)));
-        }
-        ThreadPool { workers, dispatcher: Some(dispatcher) }
-    }
-    pub fn execute<F>(&self, f: F)
-    where
-        F: FnOnce() + Send + 'static,
-    {
-        let job = Box::new(f);
-
-        self.dispatcher.as_ref().unwrap().send(job).unwrap();
-    }
-}
-
-impl Drop for ThreadPool{
-    fn drop(&mut self) {
-        drop(self.dispatcher.take());
-
-        for worker in self.workers.drain(..) {
-            println!("Shutting down worker {}", worker.id);
-
-            worker.thread.join().unwrap();
-        }
-    }
-}
-
-struct Worker {
-    id: usize,
-    thread: thread::JoinHandle<()>,
-}
-
-impl Worker {
-    fn new(id: usize, receiver: Arc<Mutex<mpsc::Receiver<Job>>>) -> Worker {
-        let thread = thread::spawn(move || loop {
-            let message = receiver.lock().expect("Worker attempted to acquire Mutex from Panicked thread, shutting down.").recv();
-
-            match message {
-                Ok(job) => {
-                    println!("Worker {} got a job; executing.", id);
-                    job();
-                },
-                Err(_) => {
-                    println!("Worker {} is shutting down.", id);
-                    break;
-                }
-            }
-        });
-        Worker { id, thread}
     }
 }
