+++ rust-web-server/src/main.rs
@@ -1,6 +1,6 @@
 use std::net::TcpListener;
 
-use rust_web_server::ThreadPool;
+use rust_web_server::multithreading::threadpool::ThreadPool;
 
 fn main() {
     let listener = TcpListener::bind("127.0.0.1:7878").unwrap();
